requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
cffi==1.14.6
cryptography==3.4.8
Django==3.2.7
django-bootstrap-v5==1.0.5
django-rest-knox==4.1.0
djangorestframework==3.12.4
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.2
